DCC XML parser
This adds a patterns.DccRecordXMLParser class for parsing DCC record XML
documents.  This interface should be more stable than the HTML, and should
require significantly fewer heuristics and lines of code.

Not all fields are supported at the moment because of limitations in the
XML that should be easily rectified by the DCC folks.