Remove (most) HTML parsing in favour of XML parsing
This commit removes the majority of the HTML parsing used previously to extract information
from DCC documents. This was fragile and reliant on regular expressions and was likely to
cause more future problems as the DCC is altered. It has been replaced by the XML parser
committed previously to this branch.

Some HTML parsing is still used to identify when an error occurs. This is needed because
the DCC will produce an HTML page in the event of an error, even if an XML document has
been requested.

In addition, internal document IDs are now recognised for "referenced by" and "reference
of" fields, though ideally these should be DCC numbers (we have requested this from the
developers).