Update code to Python 3.5 and move cookie authentication to comms.py
This commit contains fixes to the code to allow it to be interpreted by
Python 3.5. It also contains some changes to the code to move the cookie
authentication routines to comms.py and away from the command line
interpreter, prompted by some of the cookie code being incompatible
with Python 3.5.