HttpFetcher: break out method for ecp-cookie-init
This allows users of the fetcher to call this method explicitly if needed.

It's unfortunate that the authentication mechanism forces the callers to
parse returned pages to know if authentication has failed.  This should be
handled entirely inside of the fetcher without the caller having to worry
about it at all.

This new method also has some smarter error handling.
In case of ecp-cookie-init failure, klist will be called to determine if
lack of kerberos principle is the failure, and an appropriate exception will
be thrown.